# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Be Kind - Full stack restaurant and lifestyle app API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to restaurant cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{id}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # ORDERS
...
  /shop/cart/items:
    post:
      operationId: addToShopCart
      tags: [shop-cart]
      summary: Add product to shop cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToShopCartBody"
      responses:
        "200":
          description: Updated shop cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopCart"
  /shop/cart/items/{id}:
    put:
      operationId: updateShopCartItem
      tags: [shop-cart]
      summary: Update shop cart item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated shop cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopCart"
    delete:
      operationId: removeShopCartItem
      tags: [shop-cart]
      summary: Remove item from shop cart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated shop cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopCart"

  # SHOP ORDERS
...
  /events:
    get:
      operationId: getEvents
      tags: [events]
      summary: Get all events
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: upcoming
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}/register:
    post:
      operationId: registerForEvent
      tags: [events]
      summary: Register for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Registration confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventRegistration"
  /events/{id}/cancel:
    post:
      operationId: cancelEventRegistration
      tags: [events]
      summary: Cancel event registration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
    RegisterBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
          nullable: true
        acceptedTerms:
          type: boolean
      required: [firstName, lastName, email, password, phone, acceptedTerms]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          nullable: true
        loyaltyPoints:
          type: integer
        loyaltyLevel:
          type: string
        createdAt:
          type: string
      required: [id, email, firstName, lastName, phone, loyaltyPoints, loyaltyLevel, createdAt]

    # MENU
...
    CartItem:
      type: object
      properties:
        id:
          type: integer
        dishId:
          type: integer
        dishName:
          type: string
        dishPrice:
          type: number
        imageUrl:
          type: string
          nullable: true
        quantity:
          type: integer
        customizations:
          type: string
          nullable: true
        subtotal:
          type: number
      required: [id, dishId, dishName, dishPrice, quantity, subtotal]
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        deliveryCost:
          type: number
        total:
          type: number
        orderType:
          type: string
          nullable: true
      required: [items, subtotal, deliveryCost, total]
    AddToCartBody:
      type: object
      properties:
        dishId:
          type: integer
        quantity:
          type: integer
        customizations:
          type: string
          nullable: true
      required: [dishId, quantity]
    UpdateCartItemBody:
      type: object
      properties:
        quantity:
          type: integer
      required: [quantity]

    # ORDERS
...
    ShopCartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productPrice:
          type: number
        imageUrl:
          type: string
          nullable: true
        quantity:
          type: integer
        selectedVariants:
          type: object
          nullable: true
          additionalProperties: true
        subtotal:
          type: number
      required: [id, productId, productName, productPrice, quantity, subtotal]
    ShopCart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ShopCartItem"
        subtotal:
          type: number
        shippingCost:
          type: number
        total:
          type: number
      required: [items, subtotal, shippingCost, total]
    AddToShopCartBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        selectedVariants:
          type: object
          nullable: true
          additionalProperties: true
      required: [productId, quantity]

    # SHOP ORDERS
...
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
          nullable: true
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        location:
          type: string
        category:
          type: string
        price:
          type: number
        maxParticipants:
          type: integer
        currentParticipants:
          type: integer
        isFull:
          type: boolean
        isFree:
          type: boolean
        isRegistered:
          type: boolean
      required: [id, title, description, date, startTime, endTime, location, category, price, maxParticipants, currentParticipants, isFull, isFree, isRegistered]
    EventRegistration:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        eventTitle:
          type: string
        eventDate:
          type: string
        qrCode:
          type: string
        status:
          type: string
          enum: [confirmed, cancelled, attended]
        createdAt:
          type: string
      required: [id, eventId, eventTitle, eventDate, status, createdAt]

    # LOYALTY
...